+++ packages/lumx-react/src/components/thumbnail/Thumbnail.test.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { DisabledStateProvider } from '@lumx/react/utils';
 import { commonTestsSuiteRTL, SetupRenderOptions } from '@lumx/react/testing/utils';
 import { queryByClassName } from '@lumx/react/testing/utils/queries';
-import { fireEvent, render } from '@testing-library/react';
+import { fireEvent, render, screen } from '@testing-library/react';
 import { Thumbnail, ThumbnailProps } from './Thumbnail';
 
 const CLASSNAME = Thumbnail.className as string;
@@ -28,8 +28,19 @@ describe(`<${Thumbnail.displayName}>`, () => {
         },
     });
 
-    describe('disabled state', () => {
-        it('should not be clickable when disabled from context', () => {
+    describe('clickable button', () => {
+        it('should render clickable button', async () => {
+            const onClick = vi.fn();
+            const { thumbnail } = setup({ onClick, alt: 'Name' });
+            const button = screen.getByRole('button', { name: 'Name' });
+            expect(button).toBe(thumbnail);
+            expect(button).toHaveAttribute('type', 'button');
+
+            fireEvent.click(thumbnail as HTMLElement);
+            expect(onClick).toHaveBeenCalled();
+        });
+
+        it('should not render button in disabled context', () => {
             const onClick = vi.fn();
             const { thumbnail, container } = setup(
                 { onClick, 'aria-label': 'thumbnail' },
@@ -47,8 +58,21 @@ describe(`<${Thumbnail.displayName}>`, () => {
             fireEvent.click(thumbnail as HTMLElement);
             expect(onClick).not.toHaveBeenCalled();
         });
+    });
 
-        it('should have no href when disabled from context', () => {
+    describe('clickable link', () => {
+        it('should render clickable link', async () => {
+            const onClick = vi.fn((evt: any) => evt.preventDefault());
+            const { thumbnail } = setup({ linkProps: { href: '#' }, onClick, alt: 'Name' });
+            const link = screen.getByRole('link');
+            expect(link).toBe(thumbnail);
+            expect(link).toHaveAttribute('href', '#');
+
+            fireEvent.click(thumbnail as HTMLElement);
+            expect(onClick).toHaveBeenCalled();
+        });
+
+        it('should not render link in disabled context', () => {
             const { container, thumbnail } = setup(
                 { linkAs: 'a', linkProps: { href: '#' }, 'aria-label': 'thumbnail' },
                 {